+++ UnitTest_PZ/TestReducedSpace/TestReduced.cpp
@@ -35,7 +35,7 @@
 
 #include "TPZExtendGridDimension.h"
 
-#include "pzanalysis.h"
+#include "TPZAnalysis.h"
 
 #include "pzlog.h"
